Removed duplicate code and start of trivia state
