+++ Project/main/src/statemachine.c
@@ -16,7 +16,7 @@
 #include "minigame8.h"
 #include "minigame9.h"
 #include "minigame10.h"
-#include "catchThePokemon.h"
+#include "trivia.h"
 
 #include <zephyr/device.h>
 #include <zephyr/drivers/gpio.h>
@@ -27,13 +27,46 @@
 #include <stdio.h>
 
 LOG_MODULE_REGISTER(statemachine);
-#define MINIGAME_OFFSET 4
-// Setup state machine
-uint8_t trivia_ID;
 
-typedef enum {init_state, idle_state, end_game_state, exit_state, mg1_state, mg2_state, mg3_state, mg4_state,
-						mg5_state, mg6_state, mg7_state, mg8_state, mg9_state, mg10_state, ctp_state} statemachineStates;
+typedef enum {init_state, idle_state, end_game_state, exit_state, mg_state, trivia_state} statemachineStates;
 static int16_t end_time;
+
+void getMgThreads(char*** names, unsigned* amount, int mgID){
+	switch(mgID){
+		case 1:
+            getMg1Threads(names, amount);
+            break;
+        case 2:
+            getMg2Threads(names, amount);
+            break;
+        case 3:
+            getMg3Threads(names, amount);
+            break;
+        case 4:
+            getMg4Threads(names, amount);
+            break;
+        case 5:
+            getMg5Threads(names, amount);
+            break;
+        case 6:
+            getMg6Threads(names, amount);
+            break;
+        case 7:
+            getMg7Threads(names, amount);
+            break;
+        case 8:
+            getMg8Threads(names, amount);
+            break;
+        case 9:
+            getMg9Threads(names, amount);
+            break;
+        case 10:
+            getMg10Threads(names, amount);
+            break;
+        default:
+            break;
+    }
+}
 
 // State functions
 void init_stateFunction(statemachineStates* next_state) {
@@ -79,7 +112,7 @@ void init_stateFunction(statemachineStates* next_state) {
 	*next_state = idle_state;
 }
 
-void idle_stateFunction(statemachineStates* next_state) {
+void idle_stateFunction(statemachineStates* next_state, int* mgID, uint8_t* trivia_ID) {
 	LOG_INF("Idle\n");
 	char **names;
 	unsigned amount;
@@ -95,24 +128,57 @@ void idle_stateFunction(statemachineStates* next_state) {
 		LOG_INF("Going to exit state\n");
 		*next_state = exit_state;
 	}else if(ret >= 100){   		// if a minigame ID above 100 is assigned, it is a triva question, (change this when more than 100 games are made)
-		trivia_ID = ret - 100;		// internaly to the trivia game questions are labeled 0 to [however many are on the SD]
-		*next_state = mg4_state;	// keep in mind that increasing the amount of questions will influence required buffersizes, as well as the main stack
+		*trivia_ID = ret - 100;		// internaly to the trivia game questions are labeled 0 to [however many are on the SD]
+		*next_state = trivia_state;	// keep in mind that increasing the amount of questions will influence required buffersizes, as well as the main stack
 	} else {
-		*next_state = ret + MINIGAME_OFFSET;
+		*next_state = mg_state;
+		*mgID = ret;
 	}
 }
 
-void mg1_stateFunction(statemachineStates* next_state) { // Makes use of button and led
-	// Initialise state, enable and disable corresponding threads
-	LOG_INF("Minigame 1\n");
+void mg_stateFunction(statemachineStates* next_state, int mgID) {
+	LOG_INF("Minigame state\n");
 	int score = 0;
 
-	char **names;
-	unsigned amount;
-	getMg1Threads(&names, &amount);
+	char **names = NULL;
+	unsigned amount = 0;
+	getMgThreads(&names, &amount, mgID);
 	enableThreads(names, amount);
 
-	score = playMg1();
+		switch(mgID){
+		case 1:
+            score = playMg1();
+            break;
+        case 2:
+            score = playMg2();
+            break;
+        case 3:
+			score = playMg3();
+			break;
+		case 4:
+			score = playMg4();
+			break;
+		case 5:
+			score = playMg5();
+			break;
+		case 6:
+			score = playMg6();
+			break;
+		case 7:
+			score = playMg7();
+			break;
+		case 8:
+			score = playMg8();
+			break;
+		case 9:
+			score = playMg9();
+			break;
+		case 10:
+			score = playMg10();
+			break;
+		default:
+			break;
+    }
 
 	disableThreads(names, amount);
 	sd_set_score(score);
@@ -121,49 +187,13 @@ void mg1_stateFunction(statemachineStates* next_state) { // Makes use of button 
 	*next_state = idle_state;
 }
 
-void mg2_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 2\n");
+void trivia_stateFunction(statemachineStates* next_state, uint8_t trivia_ID) {
+	LOG_INF("Trivia\n");
 	int score = 0;
 
 	char **names;
 	unsigned amount;
-	getMg2Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg2();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg3_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 3\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg3Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg3();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg4_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 4\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg4Threads(&names, &amount);
+	getTriviaThreads(&names, &amount);
 	enableThreads(names, amount);
 
 	score = playMg4(trivia_ID);
@@ -171,136 +201,6 @@ void mg4_stateFunction(statemachineStates* next_state) { // Makes use of gyro an
 	disableThreads(names, amount);
 	sd_set_score(score);
 	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg5_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 5\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg5Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg5();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg6_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 6\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg6Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg6();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg7_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 7\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg7Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg7();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg8_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 8\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg8Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg8();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg9_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 9\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg9Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg9();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void mg10_stateFunction(statemachineStates* next_state) { // Makes use of gyro and buzzer
-	LOG_INF("Minigame 10\n");
-	int score = 0;
-
-	char **names;
-	unsigned amount;
-	getMg10Threads(&names, &amount);
-	enableThreads(names, amount);
-
-	score = playMg10();
-
-	disableThreads(names, amount);
-	sd_set_score(score);
-	show_mg_score(score);
-
-	*next_state = idle_state;
-}
-
-void ctp_stateFunction(statemachineStates* next_state) // Catch the Pokemon minigame
-{
-	char **names;
-	unsigned amount;
-	getCatchThePokemonThreads(&names, &amount);
-	enableThreads(names, amount);
-
-	int ret = playCatchThePokemon();
-
-	if ( ret < 0 )
-	{
-		LOG_ERR("Error in catch the pokemon state\n");
-		*next_state = exit_state;
-	}
-
-	disableThreads(names, amount);
-	sd_set_score(ret);
-	show_mg_score(ret);
 
 	*next_state = idle_state;
 }
@@ -345,6 +245,8 @@ void startStatemachine() {
 	bool statemachine_ongoing = 1;
 	//bool test = 0;
 	statemachineStates current_state = init_state;
+	int mgID = 0;
+	uint8_t trivia_ID;
 
 	while(statemachine_ongoing){
 // #ifndef CONFIG_TESTMODE
@@ -369,40 +271,13 @@ void startStatemachine() {
 				LOG_INF("Score after init: %d", sd_get_score());
 				break;
 			case idle_state:
-				idle_stateFunction(&current_state);
+				idle_stateFunction(&current_state, &mgID, &trivia_ID);
 				break;
-			case mg1_state:
-				mg1_stateFunction(&current_state);
+			case mg_state:
+				mg_stateFunction(&current_state, mgID);
 				break;
-			case mg2_state:
-				mg2_stateFunction(&current_state);
-				break;
-			case mg3_state:
-				mg3_stateFunction(&current_state);
-				break;
-			case mg4_state:
-				mg4_stateFunction(&current_state);
-				break;
-			case mg5_state:
-				mg5_stateFunction(&current_state);
-				break;
-			case mg6_state:
-				mg6_stateFunction(&current_state);
-				break;
-			case mg7_state:
-				mg7_stateFunction(&current_state);
-				break;
-			case mg8_state:
-				mg8_stateFunction(&current_state);
-				break;
-			case mg9_state:
-				mg9_stateFunction(&current_state);
-				break;
-			case mg10_state:
-				mg10_stateFunction(&current_state);
-				break;
-			case ctp_state:
-				ctp_stateFunction(&current_state);
+			case trivia_state:
+				trivia_stateFunction(&current_state, trivia_ID);
 				break;
 			case end_game_state:
 				end_game_stateFunction(&current_state);
